requirements.txt
asgiref==3.8.1
attrs==24.2.0
dj-database-url==2.2.0
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
requests==2.26.0
python-dotenv==0.21.0
gunicorn==20.1.0
whitenoise==6.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
timedelta==2020.12.3
